requirements.txt
<<<<<<< HEAD
# 盲道检测训练系统依赖包

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
ultralytics>=8.0.0

# 计算机视觉
opencv-python>=4.8.0
Pillow>=9.5.0
imageio>=2.31.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# 云服务
boto3>=1.28.0
botocore>=1.31.0
paramiko>=3.2.0

# 实验跟踪
wandb>=0.15.0
mlflow>=2.5.0
tensorboard>=2.13.0

# 超参数优化
optuna>=3.2.0
hyperopt>=0.2.7

# 配置管理
PyYAML>=6.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# 进度条和日志
tqdm>=4.65.0
rich>=13.4.0
loguru>=0.7.0

# 网络请求
requests>=2.31.0
urllib3>=2.0.0

# 图像处理增强
albumentations>=1.3.0
imgaug>=0.4.0

# 模型优化
onnx>=1.14.0
onnxruntime>=1.15.0
# tensorrt>=8.6.0  # 在Python 3.13上有兼容性问题，注释掉

# 移动端部署 (注释掉有兼容性问题的包)
# coremltools>=7.0  # 在Python 3.13上有兼容性问题
# tf2onnx>=1.15.0   # 在Python 3.13上有兼容性问题

# 数据验证
# great-expectations>=0.17.0  # 可选，注释掉减少依赖冲突
# pandera>=0.17.0  # 可选，注释掉减少依赖冲突

# 性能分析
# py-spy>=0.3.14  # 在Windows上可能有编译问题
# memory-profiler>=0.60.0  # 可选

# 测试 (可选，注释掉减少依赖冲突)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# pytest-mock>=3.11.0

# 代码质量 (可选，注释掉减少依赖冲突)
# black>=23.7.0
# flake8>=6.0.0
# isort>=5.12.0
# mypy>=1.5.0

# 文档生成 (可选，注释掉减少依赖冲突)
# sphinx>=7.1.0
# sphinx-rtd-theme>=1.3.0

# 语音播报系统
edge-tts>=6.1.0
pyttsx3>=2.90
pygame>=2.5.0
azure-cognitiveservices-speech>=1.34.0

# GUI界面
PyQt5>=5.15.0

# 其他工具
click>=8.1.0
python-dotenv>=1.0.0
psutil>=5.9.0
GPUtil>=1.4.0
=======
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
websockets
opencv-python
numpy
ultralytics
tensorflow
tensorflow-hub
torch
torchvision
python-dotenv
pytest
>>>>>>> origin/temp-merge
